requirements.txt
fastapi
uvicorn
python-dotenv
PyPDF2
pinecone
google-genai
groq
twilio
requests
openai
anthropic
python-multipart